Update Window examples
